+++ libs/queue/src/index.ts
@@ -1,0 +1,3 @@
+export { TRANSCRIBE_QUEUE } from './queue.constants';
+export { TRANSCRIBE_JOB, type TranscribeJobData } from './transcribe-job';
+export { redisConnection, type RedisConnection } from './redis';